% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanAveragePhenotypes.R
\name{cleanAveragePhenotypes}
\alias{cleanAveragePhenotypes}
\title{Reads, cleans and creates phenotype data with averaged individual data points for entries.}
\usage{
cleanAveragePhenotypes(inputFiles = NULL, metaDataFile = NULL,
  keepMetaCols = c("germplasmName"), ...)
}
\arguments{
\item{metaDataFile}{a file with the meta data columns, tab-delimited. Optional}

\item{keepMetaCols}{a vector of meta columns to keep in the data set.}

\item{...}{additional args to pass to function combinePhenoData (from phenoAnalysis)}

\item{phenoFiles}{a character vector of phenotype file(s).}
}
\value{
a data frame of the averaged phenotype dataset with trait columns, and meta columns exempt from removal.
}
\description{
Given a phenotype data file with the BrAPI style column naming and
a file with the metadata columns, reads phenotype data from one or more files,
 and then calculates averages for entries
}
\seealso{
\code{\link[phenoAnalysis]{combinePhenoData}},
\code{\link[phenoAnalysis]{extractPhenotypes}},
\code{\link[phenoAnalysis]{cleanMetaCols}},
\code{\link[phenoAnalysis]{summarizeTraits}}.
}
