% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanEmptyBlocks.R
\name{cleanEmptyBlocks}
\alias{cleanEmptyBlocks}
\title{Removes rows with completely empty trait values for a block}
\usage{
cleanEmptyBlocks(traitData, traitName = NULL)
}
\arguments{
\item{traitData}{a data frame of the trial dataset.}

\item{traitName}{the name of the trait}
}
\value{
a data frame of the dataset excluding the rows from the block compeletly with out values for the trait.
}
\description{
Given a dataset with the BrAPI style column naming, removes rows with completely empty trait values for a block
}
