% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanEmptyReps.R
\name{cleanEmptyReps}
\alias{cleanEmptyReps}
\title{Removes rows with completely empty trait values for a replication}
\usage{
cleanEmptyReps(traitData, traitName = NULL)
}
\arguments{
\item{traitData}{a data frame of the trial dataset.}

\item{traitName}{the name of the trait}
}
\value{
a data frame of the dataset excluding the rows from the replication compeletly with out values for the trait.
}
\description{
Given a dataset with the BrAPI style column naming, removes rows with completely empty trait values for a replication
}
