% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanMetaCols.R
\name{cleanMetaCols}
\alias{cleanMetaCols}
\title{Removes meta data columns from the phenotype data}
\usage{
cleanMetaCols(metaDataFile = NULL, phenoData = NULL, phenoFile = NULL,
  metaData = NULL, keepMetaCols = NULL)
}
\arguments{
\item{metaDataFile}{a file with the meta data columns, tab-delimited. Optional}

\item{phenoData}{a data frame of the phenotype dataset}

\item{phenoFile}{a name of the file with the phenotype dataset}

\item{metaData}{a vector of columns to remove.}

\item{keepMetaCols}{a vector of meta columns to keep in the data set.}
}
\value{
a data frame of the phenotype dataset with trait columns, and meta columns exempt from removal.
}
\description{
Given a dataset with the BrAPI style column naming, removes meta data columns from the phenotype data.
}
