% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combinePhenoData.R
\name{combinePhenoData}
\alias{combinePhenoData}
\title{combine phenotype data}
\usage{
combinePhenoData(
  phenoFiles = NULL,
  metaDataFile = NULL,
  keepMetaCols = NULL,
  keepAllTraits = FALSE
)
}
\arguments{
\item{phenoFiles}{A vector of the phenotype data files. Phenotype data file name must have a numeric trial id. e.g. file: phenotype_data_141.txt.}

\item{metaDataFile}{A character vector of meta data file, containing the non-trait cols in the phenotype file.}

\item{keepMetaCols}{character vector of meta data cols to keep. By default it keeps germplasmName, studyDbId, locationName, studyYear, replicate, blockNumber.}

\item{keepAllTraits}{a logical (TRUE or FALSE). By default combines only common traits.}
}
\value{
a data.frame of the combined phenotype data.
}
\description{
combines phenotype data from multiple files.
}
\seealso{
\code{\link[phenoAnalysis]{cleanMetaCols}} which this function uses
}
