% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractPhenotypes.R
\name{extractPhenotype}
\alias{extractPhenotype}
\title{Reads phenotype data from one or more files.}
\usage{
extractPhenotype(inputFiles, metaDataFile, ...)
}
\arguments{
\item{metaDataFile}{a file with the meta data columns, tab-delimited. Optional}

\item{...}{additional args to pass to function combinePhenoData (from phenoAnalysis)}

\item{phenoFiles}{a character vector of phenotype file(s).}
}
\value{
a data frame of the phenotype dataset with trait columns, and meta columns exempt from removal.
}
\description{
Given a dataset with the BrAPI style column naming, reads phenotype data from one or more files..
}
\seealso{
\code{\link[phenoAnalysis]{combinePhenoData}} which this function uses
}
