% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterMissingValues.R
\name{filterMissingValues}
\alias{filterMissingValues}
\title{Filters out columns and rows by percentage of missing value}
\usage{
filterMissingValues(
  phenoData = NULL,
  rowMissingFilter = 1,
  colMissingFilter = 1,
  traits = NULL
)
}
\arguments{
\item{phenoData}{a data frame of the phenotype dataset}

\item{rowMissingFilter}{a decimal missing values cut off to apply to rows. By default,
rows with missing values in all columns are removed.}

\item{colMissingFilter}{a decimal missing values cut off to apply to columns. By default,
columns with missing values in all rows are removed.}

\item{traits}{a vector of trait columns to filter. By default,
filters are applied to all columns.}
}
\value{
a data frame of the phenotype dataset without the columns and rows
with missing values above the given percentages.
}
\description{
Filters out columns and rows by percentage of missing value
}
