% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenoAnalysis.R
\name{getAdjMeans}
\alias{getAdjMeans}
\title{calculates the genotype adjusted means for a trait.}
\usage{
getAdjMeans(trialData, traitName = NULL, genotypeEffectType = "fixed")
}
\arguments{
\item{trialData}{a dataframe of the trial dataset, of the BrAPI structure and naming convention.}

\item{traitName}{a character vector of the trait name of interest in the dataset.}

\item{genotypeEffectType}{a character vector indicating the genotype effect type. By default, genotypes variable is considered fixed effects.}
}
\value{
a dataframe of the adjusted means for the genotypes.
}
\description{
calculates the genotype adjusted means for a trait.
}
