% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anovaTable.R
\name{getAnovaTable}
\alias{getAnovaTable}
\title{Generates the anova table}
\usage{
getAnovaTable(modelOut, tableType = NULL, traitName = NULL, ...)
}
\arguments{
\item{modelOut}{the anova model output of merModLmerTest class}

\item{tableType}{a character vector indicating type of anova table output. Defaults to 'html' for web display and 'text' for saving in a file (pass the file name to the 'out' argument to save output in a file) .}

\item{traitName}{the trait name to be used in the table title, optional.}

\item{...}{arguments to be passed to \code{\link[stargazer]{stargazer}}.}
}
\value{
a stargazer formatted anova table of chosen type, html type by default.
}
\description{
Generates the anova table
}
\seealso{
see \code{\link[stargazer]{stargazer}} to modify the table output.
}
