% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDataFrame.R
\name{getDataFrame}
\alias{getDataFrame}
\title{Reads phenotype data from a file.}
\usage{
getDataFrame(phenoFile)
}
\arguments{
\item{phenoFile}{a character vector of phenotype file.}
}
\value{
a data frame of the data from the given file.
}
\description{
reads phenotype data from a file..
}
