% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelDiagnostics.R, R/modelSummary.R
\name{getModelSummary}
\alias{getModelSummary}
\alias{getModelSummary}
\title{fits model and returns the model output summary}
\usage{
getModelSummary(trialData, traitName = NULL, genotypeEffectType = "fixed")

getModelSummary(trialData, traitName = NULL, genotypeEffectType = "fixed")
}
\arguments{
\item{trialData}{a dataframe of the trial dataset, of the BrAPI structure and naming convention.}

\item{traitName}{a character vector of the trait name of interest in the dataset.}

\item{genotypeEffectType}{a character vector indicating the genotype effect type. By default, genotypes variable is considered fixed effects.}

\item{trialData}{a dataframe of the trial dataset, of the BrAPI structure and naming convention.}

\item{traitName}{a character vector of the trait name of interest in the dataset.}

\item{genotypeEffectType}{a character vector indicating the genotype effect type. By default, genotypes variable is considered fixed effects.}
}
\value{
a summary of the model output.

a summary of the model output.
}
\description{
fits model and returns the model output summary

fits model and returns the model output summary
}
