% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNumericCols.R
\name{getNumericCols}
\alias{getNumericCols}
\title{removes non-numeric columns from a data frame}
\usage{
getNumericCols(phenoData = NULL, exception = c("germplasmName"))
}
\arguments{
\item{phenoData}{a data frame of the phenotype dataset}

\item{exception}{a vector of columns to skip checking if they are numeric.
By default, it skips the 'germplasmName' column.}
}
\value{
a data frame of the phenotype dataset with without the non-numeric variables.
}
\description{
removes non-numeric columns from a data frame
}
