% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeVariables.R
\name{mergeVariables}
\alias{mergeVariables}
\title{merges variables from 2 or more data files.}
\usage{
mergeVariables(dataFiles = NULL, ...)
}
\arguments{
\item{dataFiles}{a vector of data files}

\item{...}{args to pass to 'full_join' function from dplyr.}
}
\value{
a dataframe of the combined variables.
}
\description{
merges variables from 2 or more data files.
}
\seealso{
\code{\link[dplyr]{full_join}} which this function uses
}
