% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{models}
\alias{models}
\alias{fixedAugmentedRCBD}
\alias{randomAugmentedRCBD}
\alias{fixedCRD}
\alias{randomCRD}
\alias{fixedAlpha}
\alias{randomAlpha}
\alias{fixedRCBD}
\alias{randomRCBD}
\alias{fixedRowCol}
\title{fits linear mixed model}
\usage{
fixedAugmentedRCBD(traitData, traitName)

randomAugmentedRCBD(traitData, traitName)

fixedCRD(traitData, traitName)

randomCRD(traitData, traitName)

fixedAlpha(traitData, traitName)

randomAlpha(traitData, traitName)

fixedRCBD(traitData, traitName)

randomRCBD(traitData, traitName)

fixedRowCol(traitData, traitName)
}
\arguments{
\item{traitData}{a data frame of the trait data}

\item{traitName}{a character vector or string of the trait name of interest in the dataset.}
}
\value{
the model output, an object of class \code{\link[lmerTest]{lmer}}.
}
\description{
fits linear mixed model depending on the type of experimental design
as is found in the dataset and specification of the genotype effect type,
defaults to fixed effects.
}
