% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sortTraitData.R
\name{sortTraitData}
\alias{sortTraitData}
\title{sorts trait data by design factors for further analysis}
\usage{
sortTraitData(traitData)
}
\arguments{
\item{traitData}{a data frame of the trial dataset.}
}
\value{
a data frame of the dataset sorted by block number, replicate or both.
}
\description{
Given a dataset with the BrAPI style column naming, sorts trait data by design factors for further analysiss.
}
