% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structureGenoMeans.R
\name{structureAdjMeans}
\alias{structureAdjMeans}
\title{calculates the genotype adjusted means for a trait.}
\usage{
structureAdjMeans(modelOut, traitName)
}
\arguments{
\item{modelOut}{a lmer model output.}

\item{traitName}{a character vector or string of the trait name of interest in the dataset.}
}
\value{
a dataframe of the adjusted means for the genotypes.
}
\description{
calculates the genotype adjusted means for a trait.
}
