% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structureTraitData.R
\name{structureTraitData}
\alias{structureTraitData}
\title{structures trait data for further analysis}
\usage{
structureTraitData(trialData, traitName = NULL)
}
\arguments{
\item{trialData}{a data frame of the trial(s) dataset.}

\item{traitName}{a character vector of the trait of interest.}
}
\value{
a data frame of the trait dataset with columns 'germplasmName', 'locationName', 'studyYear', 'blockNumber', 'replicate', trait Name
}
\description{
Given a dataset with the BrAPI style column naming, it creates a trait dataset for further analysis.
}
