% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/averageTrait.R
\name{averageTrait}
\alias{averageTrait}
\title{calculates average trait values for genotypes}
\usage{
averageTrait(traitData, traitName, meansVariable = "germplasmName")
}
\arguments{
\item{traitData}{a data frame of the data set}

\item{traitName}{a character vector or string of the trait name of interest in the dataset.}
}
\value{
a dataframe of genotypes mean trait values.
}
\description{
calculates average trait values for genotypes
}
