% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAdjMeans.R
\name{getAdjMeans}
\alias{getAdjMeans}
\title{calculates the genotype adjusted means for a trait.}
\usage{
getAdjMeans(trialData = NULL, traitName = NULL, modelOut = NULL,
  genotypeEffectType = "fixed", adjMeansVariable = "germplasmName")
}
\arguments{
\item{trialData}{a dataframe of the trial dataset, of the BrAPI structure and naming convention. This argument is required if the modelOut argument is null.}

\item{traitName}{a character vector of the trait name of interest in the dataset.}

\item{modelOut}{a lmerModLmerTest or merModLmerTest class anova model output. This argument is required if the trialData argument is null.}

\item{genotypeEffectType}{a character vector indicating the genotype effect type. By default, genotypes variable is considered fixed effects.}
}
\value{
a dataframe of the adjusted means for the genotypes.
}
\description{
calculates the genotype adjusted means for a trait.
}
