% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizeTraits.R
\name{summarizeTraits}
\alias{summarizeTraits}
\title{groups by given factor and calculates summary data}
\usage{
summarizeTraits(
  phenoData = NULL,
  traits = NULL,
  groupBy = c("germplasmName"),
  summaryStat = c("mean")
)
}
\arguments{
\item{phenoData}{a data frame of the phenotype dataset}

\item{traits}{a character vector of traits summary to calculate for.
If not given, summary will be calculated for all traits in the data set.}

\item{groupBy}{a vector of the columns to group by, defaults to germplasmName.}

\item{summaryStat}{a vector of the summary statistics to calculate, defaults to mean.}
}
\value{
a data frame of the phenotype dataset with summarized by the grouping factors.
}
\description{
Groups by given factor and calculates summary data
}
